Refactoring operator '*'
Move 'Promisify()' and 'operator*()' in a separate
'eventuals/promisify.h' file plus minor refactoring.
