+++ eventuals/event-loop.cc
@@ -138,6 +138,12 @@ void EventLoop::DestructDefault() {
 
 ////////////////////////////////////////////////////////////////////////
 
+bool EventLoop::HasDefault() {
+  return loop != nullptr;
+}
+
+////////////////////////////////////////////////////////////////////////
+
 void EventLoop::ConstructDefaultAndRunForeverDetached() {
   ConstructDefault();
