+++ test/collect.cc
@@ -4,6 +4,7 @@
 #include <vector>
 
 #include "eventuals/iterate.h"
+#include "eventuals/promisify.h"
 #include "gtest/gtest.h"
 
 namespace eventuals::test {
