+++ test/concurrent/moveable.cc
@@ -5,8 +5,8 @@
 #include "eventuals/iterate.h"
 #include "eventuals/let.h"
 #include "eventuals/map.h"
+#include "eventuals/promisify.h"
 #include "test/concurrent/concurrent.h"
-#include "test/promisify-for-test.h"
 
 namespace eventuals::test {
 namespace {
@@ -34,11 +34,7 @@ TYPED_TEST(ConcurrentTypedTest, Moveable) {
           typename decltype(e())::template ErrorsFrom<void, std::tuple<>>,
           std::tuple<>>);
 
-  auto [future, k] = PromisifyForTest(e());
-
-  k.Start();
-
-  EXPECT_THAT(future.get(), this->OrderedOrUnorderedElementsAre(42));
+  EXPECT_THAT(*e(), this->OrderedOrUnorderedElementsAre(42));
 }
 
 } // namespace
