+++ test/concurrent/stream-fail.cc
@@ -3,10 +3,10 @@
 
 #include "eventuals/collect.h"
 #include "eventuals/map.h"
+#include "eventuals/promisify.h"
 #include "eventuals/stream.h"
 #include "gmock/gmock.h"
 #include "test/concurrent/concurrent.h"
-#include "test/promisify-for-test.h"
 
 namespace eventuals::test {
 namespace {
@@ -35,17 +35,8 @@ TYPED_TEST(ConcurrentTypedTest, StreamFail) {
           typename decltype(e())::template ErrorsFrom<void, std::tuple<>>,
           std::tuple<std::runtime_error>>);
 
-  auto [future, k] = PromisifyForTest(e());
-
-  k.Start();
-
   EXPECT_THAT(
-      // NOTE: capturing 'future' as a pointer because until C++20 we
-      // can't capture a "local binding" by reference and there is a
-      // bug with 'EXPECT_THAT' that forces our lambda to be const so
-      // if we capture it by copy we can't call 'get()' because that
-      // is a non-const function.
-      [future = &future]() { future->get(); },
+      [&]() { *e(); },
       ThrowsMessage<std::runtime_error>(StrEq("error")));
 }
