+++ test/grpc/death-client.cc
@@ -1,4 +1,5 @@
 #include "eventuals/grpc/client.h"
+#include "eventuals/promisify.h"
 #include "eventuals/then.h"
 #include "examples/protos/helloworld.grpc.pb.h"
 #include "test/grpc/death-constants.h"
