+++ test/grpc/death-server.cc
@@ -1,5 +1,6 @@
 #include "eventuals/grpc/server.h"
 #include "eventuals/head.h"
+#include "eventuals/promisify.h"
 #include "eventuals/then.h"
 #include "examples/protos/helloworld.grpc.pb.h"
 #include "test/grpc/death-constants.h"
