+++ test/grpc/server-death-test.cc
@@ -8,6 +8,7 @@
 #include "eventuals/grpc/client.h"
 #include "eventuals/grpc/server.h"
 #include "eventuals/let.h"
+#include "eventuals/promisify.h"
 #include "eventuals/terminal.h"
 #include "eventuals/then.h"
 #include "examples/protos/helloworld.grpc.pb.h"
