+++ test/iterate.cc
@@ -13,6 +13,7 @@
 
 #include "eventuals/loop.h"
 #include "eventuals/map.h"
+#include "eventuals/promisify.h"
 #include "eventuals/reduce.h"
 #include "eventuals/then.h"
 #include "gtest/gtest.h"
