+++ test/promisify-for-test.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "eventuals/scheduler.h"
+#include "eventuals/promisify.h"
 #include "test/generate-test-task-name.h"
 
 // Helper that injects a name for 'Promisify()' to simplify call sites.
