+++ test/range.cc
@@ -3,6 +3,7 @@
 #include <vector>
 
 #include "eventuals/collect.h"
+#include "eventuals/promisify.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
