+++ test/take.cc
@@ -5,6 +5,7 @@
 #include "eventuals/collect.h"
 #include "eventuals/filter.h"
 #include "eventuals/iterate.h"
+#include "eventuals/promisify.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
