+++ test/unpack.cc
@@ -3,6 +3,7 @@
 #include <string>
 
 #include "eventuals/just.h"
+#include "eventuals/promisify.h"
 #include "eventuals/then.h"
 #include "gtest/gtest.h"
